refactor(DataTable): move the logic of columns dnd inside new hook
